CID-1116483 cidr to netmask bitshifts guarded with casts
 and simple test

+++ utils/src/main/java/com/cloud/utils/net/NetUtils.java
@@ -921,7 +921,7 @@ public class NetUtils {
     }
 
     public static String getCidrNetmask(final long cidrSize) {
-        final long numericNetmask = 0xffffffff >> MAX_CIDR - cidrSize << MAX_CIDR - cidrSize;
+        final long numericNetmask = (long)0xffffffff >> MAX_CIDR - cidrSize << MAX_CIDR - cidrSize;
         return long2Ip(numericNetmask);
     }
